# openapi/fastapi.json
{
  "openapi": "3.1.0",
  "info": {
    "title": "Fuel-Aware Backend",
    "version": "0.1.0",
    "summary": "Fuel-Aware backend APIs",
    "description": "Operational endpoints for the Fuel-Aware backend across health and telemetry verification flows."
  },
  "paths": {
    "/health": {
      "get": {
        "tags": [
          "System"
        ],
        "summary": "Check FastAPI backend health",
        "description": "Returns the FastAPI runtime state and startup database reachability.",
        "operationId": "health_health_get",
        "responses": {
          "200": {
            "description": "FastAPI runtime health payload.",
            "content": {
              "application/json": {
                "schema": {
                  "additionalProperties": {
                    "type": "string"
                  },
                  "type": "object",
                  "title": "Response Health Health Get"
                }
              }
            }
          }
        }
      }
    },
    "/telemetry/sample": {
      "get": {
        "tags": [
          "Telemetry"
        ],
        "summary": "Fetch a telemetry sample",
        "description": "Returns a limited sample from the telemetry table for development-time connectivity and payload verification.",
        "operationId": "telemetry_sample_telemetry_sample_get",
        "parameters": [
          {
            "name": "limit",
            "in": "query",
            "required": false,
            "schema": {
              "type": "integer",
              "maximum": 100,
              "minimum": 1,
              "default": 10,
              "title": "Limit"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "A telemetry sample payload.",
            "content": {
              "application/json": {
                "schema": {
                  "additionalProperties": true,
                  "type": "object",
                  "title": "Response Telemetry Sample Telemetry Sample Get"
                }
              }
            }
          },
          "500": {
            "description": "The database query failed."
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    }
  },
  "components": {
    "schemas": {
      "HTTPValidationError": {
        "properties": {
          "detail": {
            "items": {
              "$ref": "#/components/schemas/ValidationError"
            },
            "type": "array",
            "title": "Detail"
          }
        },
        "type": "object",
        "title": "HTTPValidationError"
      },
      "ValidationError": {
        "properties": {
          "loc": {
            "items": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "integer"
                }
              ]
            },
            "type": "array",
            "title": "Location"
          },
          "msg": {
            "type": "string",
            "title": "Message"
          },
          "type": {
            "type": "string",
            "title": "Error Type"
          }
        },
        "type": "object",
        "required": [
          "loc",
          "msg",
          "type"
        ],
        "title": "ValidationError"
      }
    }
  }
}